Put site update read counter in `login` instead
