+++ weasyl/test/db_utils.py
@@ -72,10 +72,6 @@ def create_user(full_name="", birthday=None, config=None,
     if email_addr is not None:
         d.engine.execute("UPDATE login SET email = %(email)s WHERE userid = %(id)s",
                          email=email_addr, id=user.userid)
-    # Create the latest-read site update counter for this user
-    latest_updateid = d.engine.scalar("SELECT MAX(updateid) FROM siteupdate")
-    d.engine.execute("INSERT INTO siteupdateread (userid, updateid) VALUES (%(userid)s, %(updateid)s)",
-                     userid=user.userid, updateid=latest_updateid)
     # Force the userID to a user-defined value and return it
     if user_id is not None:
         d.engine.execute("UPDATE login SET userid = %(newid)s WHERE userid = %(oldid)s", newid=user_id, oldid=user.userid)
